fix redirect at start
