+++ pages/index.tsx
@@ -1,27 +1,18 @@
-import { User } from 'firebase/auth';
-import type { NextPage } from 'next'
-import { AppProps } from 'next/app'
-import { useRouter } from 'next/router';
-import Auth from '../components/Auth';
-import Dashboard from '../components/overview/Dashboard';
-import Loading from '../components/Loading';
-import { useUserContext } from '../context/userContext';
+import { useRouter } from "next/router";
+import Auth from "../components/Auth";
+import { useUserContext } from "../context/userContext";
 
 const Home = () => {
-  const { user, loading, error }: any = useUserContext();
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  const { user }: any = useUserContext();
+  // eslint-disable-next-line react-hooks/rules-of-hooks
   const router = useRouter();
 
-  return (
-    <div>
-      {
-        loading ?
-          <Loading/> :
-          user && !error ?
-            <Dashboard /> :
-            null
-      }
-    </div>
-  )
+  if (user) {
+      router.push("/owesme");
+  }
+
+  return <Auth />
 }
 
 export default Home
